+++ packages/calcite-components/src/components/table-header/table-header.tsx
@@ -57,6 +57,9 @@ export class TableHeader implements LocalizedComponent, LoadableComponent, T9nCo
   @Prop() numberCell = false;
 
   /** @internal */
+  @Prop() parentRowAlignment: Alignment = "start";
+
+  /** @internal */
   @Prop() parentRowIsSelected: boolean;
 
   /** @internal */
@@ -223,12 +226,15 @@ export class TableHeader implements LocalizedComponent, LoadableComponent, T9nCo
           class={{
             [CSS.bodyRow]: this.parentRowType === "body",
             [CSS.footerRow]: this.parentRowType === "foot",
+            [CSS.contentCell]: !this.numberCell && !this.selectionCell,
             [CSS.numberCell]: this.numberCell,
             [CSS.selectionCell]: this.selectionCell,
             [CSS.selectedCell]: this.parentRowIsSelected,
             [CSS.multipleSelectionCell]: this.selectionMode === "multiple",
             [CSS.staticCell]: staticCell,
             [CSS.lastCell]: this.lastCell && (!this.rowSpan || (this.colSpan && !!this.rowSpan)),
+            [this.parentRowAlignment]:
+              this.parentRowAlignment === "center" || this.parentRowAlignment === "end",
           }}
           colSpan={this.colSpan}
           onBlur={this.onContainerBlur}
@@ -249,13 +255,15 @@ export class TableHeader implements LocalizedComponent, LoadableComponent, T9nCo
               scale={getIconScale(this.scale)}
             />
           )}
-          <span
-            aria-hidden={true}
-            aria-live={this.focused ? "polite" : "off"}
-            class={CSS.assistiveText}
-          >
-            {(this.selectionCell || this.numberCell) && this.screenReaderText}
-          </span>
+          {(this.selectionCell || this.numberCell) && (
+            <span
+              aria-hidden={true}
+              aria-live={this.focused ? "polite" : "off"}
+              class={CSS.assistiveText}
+            >
+              {this.screenReaderText}
+            </span>
+          )}
         </th>
       </Host>
     );